Moved all compiler-related functions into one module.
