+++ svl/__init__.py
@@ -1,7 +1,3 @@
-from .svl import parse_svl
-
-__all__ = ["parse_svl"]
-
 from ._version import get_versions
 
 __version__ = get_versions()["version"]
